+++ tests/test_vector_data_importer/test_conversion.py
@@ -1,5 +1,6 @@
 import json
 
+import numpy as np
 import pytest
 import shapely
 from shapely.testing import assert_geometries_equal
@@ -16,19 +17,18 @@ from .utils import *
 
 
 def get_schematisation_layers(target_gpkg, target_object):
+    temp_gpkg = str(get_temp_copy(target_gpkg))
     return {
-        "structure_layer": gpkg_layer(target_gpkg, target_object),
-        "channel_layer": gpkg_layer(target_gpkg, "channel"),
-        "node_layer": gpkg_layer(target_gpkg, "connection_node"),
-        "cross_section_location_layer": gpkg_layer(
-            target_gpkg, "cross_section_location"
-        ),
+        "structure_layer": gpkg_layer(temp_gpkg, target_object),
+        "channel_layer": gpkg_layer(temp_gpkg, "channel"),
+        "node_layer": gpkg_layer(temp_gpkg, "connection_node"),
+        "cross_section_location_layer": gpkg_layer(temp_gpkg, "cross_section_location"),
     }
 
 
 def get_source_layer(name, layername):
-    src = SOURCE_PATH.joinpath(name).with_suffix(".gpkg")
-    return gpkg_layer(str(src), layername)
+    src = str(get_temp_copy(SOURCE_PATH.joinpath(name).with_suffix(".gpkg")))
+    return gpkg_layer(src, layername)
 
 
 def get_import_config_path(import_config_name):
@@ -55,8 +55,8 @@ def compare_layer(layer, ref_layer):
 
 
 def compare_results(ref_name, layers, target_object):
-    ref_gpkg = DATA_PATH.joinpath("ref", ref_name).with_suffix(".gpkg")
-    ref_layers = get_schematisation_layers(ref_gpkg, target_object)
+    src = DATA_PATH.joinpath("ref", ref_name).with_suffix(".gpkg")
+    ref_layers = get_schematisation_layers(src, target_object)
     # check attributes: id and geom - anything else is hopefully covered by unit tests
     for name, layer in layers.items():
         compare_layer(layer, ref_layers[name])
@@ -68,7 +68,7 @@ def test_multi_import(qgis_application):
     layer_pt = get_source_layer("culvert_2layers", "culvert_point")
     layer_line = get_source_layer("culvert_2layers", "culvert_line")
     import_config = get_import_config_path("culvert")
-    target_gpkg = get_schematisation_copy("test_2culverts", "test_multi_import.gpkg")
+    target_gpkg = SCHEMATISATION_PATH.joinpath("test_2culverts.gpkg")
     layers = get_schematisation_layers(target_gpkg, "culvert")
     importer = CulvertsImporter(layer_pt, target_gpkg, import_config, **layers)
     importer.import_features()
@@ -81,9 +81,7 @@ def test_integrate_weir_too_long(qgis_application):
     # Test integrating multiple weirs on a channel where the total length of the weirs
     # is larger than the channel
     import_config = get_import_config_path("integrate_weirs_nosnap_too_long.json")
-    target_gpkg = get_schematisation_copy(
-        "schematisation_channel.gpkg", "test_weirs_too_long.gpkg"
-    )
+    target_gpkg = SCHEMATISATION_PATH.joinpath("schematisation_channel.gpkg")
     src_layer = get_source_layer("weirs_too_long.gpkg", "dhydro_weir")
     layers = get_schematisation_layers(target_gpkg, "weir")
     importer = WeirsImporter(src_layer, target_gpkg, import_config, **layers)
@@ -95,9 +93,7 @@ def test_integrate_weir_too_long(qgis_application):
 
 def test_integrate_isolated_weir(qgis_application):
     import_config = get_import_config_path("integrate_weirs_snap.json")
-    target_gpkg = get_schematisation_copy(
-        "schematisation_channel_with_weir.gpkg", "test_weir_isolated.gpkg"
-    )
+    target_gpkg = SCHEMATISATION_PATH.joinpath("schematisation_channel_with_weir.gpkg")
     src_layer = get_source_layer("weir_isolated.gpkg", "dhydro_weir")
     layers = get_schematisation_layers(target_gpkg, "weir")
     importer = WeirsImporter(src_layer, target_gpkg, import_config, **layers)
@@ -107,9 +103,10 @@ def test_integrate_isolated_weir(qgis_application):
 
 def test_import_connection_nodes(qgis_application):
     import_config = get_import_config_path("import_connection_nodes.json")
-    target_gpkg = get_schematisation_copy(
-        "schematisation_channel.gpkg", "test_connection_nodes.gpkg"
+    target_gpkg = get_temp_copy(
+        SCHEMATISATION_PATH.joinpath("schematisation_channel.gpkg")
     )
+
     src_layer = get_source_layer("connection_nodes.gpkg", "connection_nodes")
     target_layer = gpkg_layer(target_gpkg, "connection_node")
     importer = ConnectionNodesImporter(
@@ -117,7 +114,7 @@ def test_import_connection_nodes(qgis_application):
     )
     importer.import_features()
     ref_layer = gpkg_layer(
-        DATA_PATH.joinpath("ref", "test_import_connection_nodes.gpkg"),
+        get_temp_copy(DATA_PATH.joinpath("ref", "test_import_connection_nodes.gpkg")),
         "connection_node",
     )
     compare_layer(target_layer, ref_layer)
@@ -131,10 +128,18 @@ def test_import_weirs(qgis_application, integrate: bool, snap: bool):
     test_name = f"{'integrate' if integrate else 'import'}_weirs_{'no' if snap else ''}snap.json"
     import_config = get_import_config_path(test_name)
     src_layer = get_source_layer("weirs.gpkg", "dhydro_weir")
-    target_gpkg = get_schematisation_copy(
-        "schematisation_channel_with_weir.gpkg", f"test_{test_name}.gpkg"
-    )
+    target_gpkg = SCHEMATISATION_PATH.joinpath("schematisation_channel_with_weir.gpkg")
     layers = get_schematisation_layers(target_gpkg, target_layer_name)
     importer = WeirsImporter(src_layer, target_gpkg, import_config, **layers)
     importer.import_features()
     compare_results(f"test_{test_name}", layers, target_layer_name)
+
+
+def test_fix_positioning(qgis_application):
+    import_config = get_import_config_path("import_weirs_fix.json")
+    src_layer = get_source_layer("weirs_fix_positions.gpkg", "dhydro_weir")
+    target_gpkg = SCHEMATISATION_PATH.joinpath("schematisation_channel.gpkg")
+    layers = get_schematisation_layers(target_gpkg, "weir")
+    importer = WeirsImporter(src_layer, target_gpkg, import_config, **layers)
+    importer.import_features()
+    compare_results("test_weirs_fix_positions", layers, "weir")