+++ HISTORY.rst
@@ -4,7 +4,7 @@ History
 2.3.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Improve handling of structures integrated on channels: prevent structures from sticking out on the start or end of the channel; prevent overlapping structures; prevent too short channels in between structures and at the start or end of the channel.
 
 
 2.3.1 (2025-07-16)